# .well-known/openapi.yaml
openapi: 3.0.1
info:
    title: 智心铸灵-异兽终端
    description: 该插件是“智心铸灵”项目的辅助插件，功能包括生成一只异兽（一种奇异的生物）、与异兽交互、显示异兽的信息。
    version: "v1"
servers:
    - url: http://139.224.200.123:8088
paths:
    /generate_pet:
        post:
            operationId: generatePet
            summary: 当用户要求获取一只异兽时，生成一只奇异的生物，它属于一种独特的种族，拥有自己的名字、性格等信息。
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/generatePet"
            responses:
                "200":
                    description: 成功生成异兽
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/messageResponse"
    /get_information:
        post:
            operationId: getInformation
            summary: 展示异兽信息。
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/getInformation"
            responses:
                "200":
                    description: 成功展示信息！
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/messageResponse"
components:
    schemas:
        generatePet:
            type: object
            required: [query]
            properties:
                query:
                    type: string
                    description: 用户请求开始游戏。
        getInformation:
            type: object
            required: [query]
            properties:
                query:
                    type: string
                    description: 用户请求获取异兽信息
        messageResponse:
            type: object
            required: [message]
            properties:
                message:
                    type: string
                    description: 回复信息